Added slackObj.attachments to slackBotIn changed payload to message.getBody()
Some bots (the IFTTT bot for example) send messages contained in the attachment part of the message and do not have a message.text element.  The slack-client library has a function getBody() that will pull the text body out of the attachment.  
Updated the slackBotIn function to set the payload to message.getBody()  rather than message.text
Added the attachment object to the slackObj